Extract update session params func
